Demolevel1
Or actual level 1. Depends if it works.
Added some of the newer sprites (like background trees which will not collie with player) and made particles for the level. I placed old portal in the level to show where the new one could go, and also places those "coins"  to show where they would be. Oh right, the spring is there too.